Ref: zts and its helpers.
