--- src/zts-types.ts
@@ -1,40 +1,0 @@
-/**
- * This file is based on https://github.com/sachinraja/zod-to-ts
- *
- * MIT License
- *
- * Copyright (c) 2021 Sachin Raja
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-
-import ts from "typescript";
-import { z } from "zod";
-import { HandlingVariant, SchemaHandler } from "./schema-walker";
-
-export type LiteralType = string | number | boolean;
-
-export type ZTSOptions = {};
-
-export type ZTSContext = {} & ZTSOptions;
-
-export type Producer<
-  T extends z.ZodTypeAny,
-  Variant extends HandlingVariant = "regular"
-> = SchemaHandler<T, ts.TypeNode, ZTSContext, Variant>;
